% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-hurricane.R
\name{compute_StatHurricane}
\alias{compute_StatHurricane}
\title{compute_group function for StatHurricane}
\usage{
compute_StatHurricane(data, scales)
}
\description{
compute_group function for StatHurricane
}
