% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomhurricane.R
\name{create_points}
\alias{create_points}
\title{Creates a data.frame with points having (long, lat) as center, and the given wind radius.
Points are created between the range of start and end degrees, scaled by the scale factor.
This calculation uses geosphere::destPoint. Distance is scaled using 1852 and the scale factor.}
\usage{
create_points(long, lat, start_deg, end_deg, radius, clr, fl, scale_factor)
}
\arguments{
\item{long}{Longitude}

\item{lat}{Latitude}

\item{start_deg}{Starting degree}

\item{end_deg}{Ending degree}

\item{radius}{Maximum wind radius along that quadrant}

\item{clr}{The color value to be used.}

\item{scale_factor}{The scaling factor for recalibrating the distance}
}
\value{
A data.frame consisting of x (longitude), y (latitude), colour, fill
}
\description{
Creates a data.frame with points having (long, lat) as center, and the given wind radius.
Points are created between the range of start and end degrees, scaled by the scale factor.
This calculation uses geosphere::destPoint. Distance is scaled using 1852 and the scale factor.
}
\examples{
\dontrun{
}
}
