% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hurricana_Project_Geom_Data.R
\name{data_fetch}
\alias{data_fetch}
\title{PART 1
Read the data}
\usage{
data_fetch(filename = "atlc_1988_2015.txt", dir = "./Data")
}
\arguments{
\item{filename}{A character string with the full file name}

\item{dir}{A characther string with path to data file}
}
\value{
The function returns a data table with EBTRK data.
}
\description{
This function reads a rectangular data file and returns a tibble to
fetch \code{data_tidy}
}
