% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hurricana_Project_Geom_Data.R
\name{data_filter_hurricane}
\alias{data_filter_hurricane}
\title{PART 3
Select data by hurricane}
\usage{
data_filter_hurricane(data, name = "IKE", year = 2008)
}
\arguments{
\item{data}{A tibble resulted from \code{data_tidy}}

\item{name}{Hurricane's name}

\item{year}{Year of event's occurrence}
}
\value{
Hurricane data
}
\description{
This function takes the clean data produced by the \code{data_tidy}
}
\examples{
\dontrun{
    data_filter_hurricane(data, name = "KATRINA", year = "2005")
    data_filter_hurricane(data, name = "IKE", year = "2008")
}

}
