% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hurricana_Project_Geom_Data.R
\name{data_tidy}
\alias{data_tidy}
\title{To tide the data}
\usage{
data_tidy(data)
}
\arguments{
\item{data}{Tibble that is created using the \code{data_fetch} function}
}
\value{
Returns a "tidy" tibble
}
\description{
This function takes the data from \code{data_fetch}
}
