% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurricane_data.R
\name{fetch_data}
\alias{fetch_data}
\title{Read in Extended Best Track (EBTRK) data}
\usage{
fetch_data(filename = "ebtrk_atlc_1988_2015.txt", dir = "./data")
}
\arguments{
\item{filename}{A character string with the full file name}

\item{dir}{A characther string with path to data file}
}
\value{
The function returns a data table with EBTRK data.

Returns a tibble with EBTRK hurricane data
}
\description{
This function reads in raw EBTRK rectangular data and returns a tibble
}
