% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomhurricane.R
\name{format_hurricane_data}
\alias{format_hurricane_data}
\title{Creates a tribble with a row for for each wind radii per wind direction}
\usage{
format_hurricane_data(df)
}
\arguments{
\item{df}{The tribble containing a row of hurricane data}
}
\value{
A trible consisting of storm_id, datetime, longitude, latitude, wind_speed, ne, nw, se, sw
}
\description{
Creates a tribble with a row for for each wind radii per wind direction
}
\examples{
\dontrun{
}
}
