% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurricane_geom.R
\name{geom_hurricane}
\alias{geom_hurricane}
\title{Geom hurricane build layer}
\usage{
geom_hurricane(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{mapping through ggplot2}

\item{data}{map data through ggplot2}

\item{stat}{map statistic through ggplot2}

\item{position}{map position through ggplot2}

\item{na.rm}{remove NAs}

\item{show.legend}{default to layer}

\item{inherit.aes}{inherit aes from main ggplot layer}

\item{...}{more arguments for the layer}
}
\value{
returns a ggplot2 graphical object
}
\description{
This function builds a geom layer based on the geom specification in \code{GeomHurricane}
}
