% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hurricana_Project_Geom.R
\name{GeomHurricane}
\alias{GeomHurricane}
\title{PART 2
Create GeomHurricane}
\usage{
GeomHurricane(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)

GeomHurricane(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{mapping through ggplot2}

\item{data}{map data through ggplot2}

\item{stat}{map statistic through ggplot2}

\item{position}{map position through ggplot2}

\item{na.rm}{remove NAs}

\item{show.legend}{default to layer}

\item{inherit.aes}{inherit aes from main ggplot layer}

\item{...}{more arguments for the layer}
}
\value{
returns a ggplot2 graphical object
}
\description{
This function creates the geom

This function builds a geom layer based on \code{GeomHurricane}
}
