% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurricane.R
\name{hurricane}
\alias{hurricane}
\title{Extract single hurricane from tidied ext_tracks}
\usage{
hurricane(.data, name, .date = NULL)
}
\arguments{
\item{.data}{Tidied ext_tracks data.frame}

\item{name}{Name (ID) of the hurricane as stored in .data}

\item{.date}{date-time string, such as "2008-09". The function will filter
with \code{stringr::str_detect}}
}
\value{
A single observation of a hurricane in a data.frame
}
\description{
Extract single hurricane from tidied ext_tracks
}
