% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hurricana_Project_Geom.R
\name{projections}
\alias{projections}
\title{PART 1
We are working with polygon point projections
Helps to compute geocode projections by quadrant}
\usage{
projections(point, quadrant, distance)
}
\arguments{
\item{point}{longitude and latitude}

\item{quadrant}{one of 4 quadrants}

\item{distance}{scaled distance in meters for the projection}
}
\value{
data frame with point projections
}
\description{
PART 1
We are working with polygon point projections
Helps to compute geocode projections by quadrant
}
