% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomhurricane.R
\name{read_hurricane_data}
\alias{read_hurricane_data}
\title{Creates a tribble with a row for for each wind radii per wind direction
based on the given file for the storm name and the specific time.
Longitude is negated to -longitude. Year, month, day and hour are combined into a single
datetime column of format yyyy-mm-dd hh:00:00}
\usage{
read_hurricane_data(path, storm_fullname, exacttime)
}
\arguments{
\item{path}{File path containing the storm data.}

\item{storm_fullname}{This should be of the format <stormid>-<year> Eg: IKE-2008}

\item{exacttime}{Complete date / time value to restrict the storm to a single row. Eg: 2008-09-13 06:00:00}
}
\value{
A tribble consisting of storm_id, datetime, longitude, latitude, wind_speed, ne, nw, se, sw
}
\description{
Creates a tribble with a row for for each wind radii per wind direction
based on the given file for the storm name and the specific time.
Longitude is negated to -longitude. Year, month, day and hour are combined into a single
datetime column of format yyyy-mm-dd hh:00:00
}
\examples{
\dontrun{
ike_data <- read_hurricane_data(paste0(getwd(), "/ebtrk_atlc_1988_2015.txt"), "IKE-2008",
 "2008-09-13 06:00:00")
}
}
