% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-hurricane.R
\name{stat_hurricane}
\alias{stat_hurricane}
\title{export}
\usage{
stat_hurricane(mapping = NULL, data = NULL, geom = "polygon",
  position = "identity", show.legend = NA, inherit.aes = TRUE, ...)
}
\description{
export
}
