% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurricane_data.R
\name{tidy_data}
\alias{tidy_data}
\title{Tidy EBTRK hurricane data}
\usage{
tidy_data(data)
}
\arguments{
\item{data}{Tibble that is created using the \code{fetch_data} function}
}
\value{
Returns a "tidy" tibble
}
\description{
This function takes the data fetched by the \code{fetch_data} function and "tidys" up the data
}
