% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-ext-tracks.R
\name{tidy_ext_tracks}
\alias{tidy_ext_tracks}
\title{Tidy ext_tracks data}
\usage{
tidy_ext_tracks(ext_tracks)
}
\arguments{
\item{ext_tracks}{ext_tracks data frame object}
}
\value{
A tidied version of ext_tracks as recommended by the authors of the
 Coursera course
}
\description{
Tidy ext_tracks data
}
