% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_rose.R
\name{wind_rose_piece}
\alias{wind_rose_piece}
\title{Coordinates of piece of wind rose}
\usage{
wind_rose_piece(lng, lat, radius, piece_key)
}
\arguments{
\item{lng}{Longitude of center}

\item{lat}{Latitude of center}

\item{radius}{The radius of the piece}

\item{piece_key}{A string of either c("ne", "se", "sw", "nw")}
}
\value{
Latitude and Longitude coordinates for a piece of a wind rose
}
\description{
Coordinates of piece of wind rose
}
