% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurricane_data.R
\name{data_filter_hurricane}
\alias{data_filter_hurricane}
\title{Filter out data by hurricane}
\usage{
data_filter_hurricane(data, name = "alberto", year = 2012)
}
\arguments{
\item{data}{A tibble with tidy hurricane data}

\item{name}{A character string of the name of the hurricane}

\item{year}{And integer or character string with year of the storm -- as the same named storm can happen in multiple years}
}
\value{
Filtered hurricane data
}
\description{
This function takes the clean data produced by the \code{data_tidy} function and produces a new filtered dataset
}
\note{
The year of the storm is also needed because storm names are recycled
}
\examples{
\dontrun{
    data_filter_hurricane(data, name = "alberto", year = 1988)
    data_filter_hurricane(data, name = "alberto", year = "2006")
    data_filter_hurricane(data, name = "alberto", year = 2012)
}

}
