% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurricane_data.R
\name{data_tidy}
\alias{data_tidy}
\title{Tidy EBTRK hurricane data}
\usage{
data_tidy(data)
}
\arguments{
\item{data}{Tibble that is created using the \code{data_fetch} function}
}
\value{
Returns a "tidy" tibble
}
\description{
This function takes the data fetched by the \code{data_fetch} function and "tidys" up the data
}
