% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomhurricane.R
\name{geom_hurricane}
\alias{geom_hurricane}
\title{Create an instance of GeomHurricane
source from geom_polygon}
\usage{
geom_hurricane(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{data}{The hurricane data}

\item{stat}{Statistics object currently not doing any transformations}

\item{position}{Origin position}
}
\value{
A layer for plotting the hurricane data.
}
\description{
Create an instance of GeomHurricane
source from geom_polygon
}
