% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-hurricane.R
\name{geom_hurricane}
\alias{geom_hurricane}
\title{Create a windrose of a storm}
\usage{
geom_hurricane(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\description{
Create a windrose of a storm
}
\section{Aesthetics}{

Required aesthetics are bold
\itemize{
 \item \strong{x}
 \item \strong{y}
 \item \strong{r_ne}
 \item \strong{r_se}
 \item \strong{r_sw}
 \item \strong{r_nw}
 \item fill
 \item scale_radii
}
}

