% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurricane_geom.R
\docType{data}
\name{GeomHurricane}
\alias{GeomHurricane}
\title{Create GeomHurricane}
\format{
An object of class \code{GeomHurricane} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomHurricane
}
\description{
This function creates a class with data manipulations to create geoms
}
\keyword{datasets}
