% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurricane_geom.R
\name{projections}
\alias{projections}
\title{Create polygon point projections}
\usage{
projections(point, quadrant, distance)
}
\arguments{
\item{point}{starting longitude and latitude of the projection}

\item{quadrant}{one of 4 quadrants (in arc) for the projection}

\item{distance}{scaled distance in meters for the projection}
}
\value{
data frame with point projections
}
\description{
Helper function to create polygon point projections
}
